+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2025-07-31
+
 ### Fixed
 
 - Adapted the build so that README.md does correctly include usage helps when updated during the release process